data-analysis/requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.40.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Visualization
plotly>=5.18.0
kaleido>=0.2.1

# Config
pyyaml>=6.0.1

# CLI
rich>=13.7.0

# Notebook
jupyter>=1.0.0
